Prompt toolkit req added; hello world started
dev-requirements now includes the main requirements: should be able to
install everything in one command.

A basic hello world dialog added to main,
where we'll focus our efforts.

Code style in Main indicates the project requirements,
including return types,
docstrings,
and proper spacing.

Black will take care of most of it: for others, we have to follow
the flake8 errors.

- side note: adjusted pre-commit config to use the remote hooks
for flake8 and isort. Found that actually works, compared to
calling `python` directly in a local hook
(which can fail on some systems if the `python` command is pointing to
the wrong interpreter